refactor: Modularize Snake's game logic
